fix state transitions and redraw loops
