+++ ui/analyse/src/view/moveView.ts
@@ -22,7 +22,7 @@ export const renderIndex = (ply: Ply, withDots?: boolean): VNode =>
 
 export function renderMove(ctx: Ctx, node: Tree.Node): VNode[] {
   const ev = cevalView.getBestEval({ client: node.ceval, server: node.eval });
-  const nodes = [h('san', { attrs: { 'aria-live': 'polite' } }, fixCrazySan(node.san!))];
+  const nodes = [h('san', fixCrazySan(node.san!))];
   if (node.glyphs && ctx.showGlyphs) node.glyphs.forEach(g => nodes.push(renderGlyph(g)));
   if (node.shapes?.length) nodes.push(h('shapes'));
   if (ev && ctx.showEval) {
