+++ ui/round/css/_nvui.scss
@@ -7,11 +7,6 @@
 }
 
 .nvui {
-  label,
-  h6 {
-    pointer-events: none;
-  }
-
   h2 {
     font-weight: bold;
     margin-top: 0.5em;
